# swagger.yaml
openapi: 3.0.3
info:
  title: Dental Clinic API
  version: 1.0.0
  description: Comprehensive API contract for all `/api/v1` routes.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /invoices/{invoice}/items:
    get: { tags: [Billing], summary: List invoice items, parameters: [{ $ref: '#/components/parameters/InvoiceId' }], responses: { '200': { $ref: '#/components/responses/Ok' } } }
    post: { tags: [Billing], summary: Create invoice item, parameters: [{ $ref: '#/components/parameters/InvoiceId' }], responses: { '201': { $ref: '#/components/responses/Created' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  responses:
    Ok:
      description: Successful response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSuccess'
    Created:
      description: Resource created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSuccess'
...
  schemas:
    ApiSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, nullable: true }
        data: { type: object, nullable: true, additionalProperties: true }
        meta: { type: object, nullable: true, additionalProperties: true }
...
  parameters:
...
    InvoiceId: { name: invoice, in: path, required: true, schema: { type: integer } }
...